respond to some of the review
